backend/test_api.py
"""
Test script to verify API functionality.
Run this after starting the server with: python -m app.main
"""
import requests
import json
from time import sleep

BASE_URL = "http://localhost:8000"

def print_response(title, response):
    """Pretty print API response."""
    print(f"\n{'='*60}")
    print(f"🧪 {title}")
    print(f"{'='*60}")
    print(f"Status: {response.status_code}")
    try:
        print(f"Response: {json.dumps(response.json(), indent=2, ensure_ascii=False)}")
    except:
        print(f"Response: {response.text}")
    print()


def test_api():
    """Run a series of API tests."""
    
    print("\n🚀 TeleCluster Orchestrator API - Test Suite")
    print("="*60)
    
    # 1. Health Check
    print("\n1️⃣  Testing Health Check...")
    r = requests.get(f"{BASE_URL}/health")
    print_response("Health Check", r)
    assert r.status_code == 200, "Health check failed"
    
    # 2. Register new user
    print("\n2️⃣  Testing User Registration...")
    register_data = {
        "email": "testuser@pucp.edu.pe",
        "password": "password123",
        "full_name": "Usuario de Prueba",
        "role": "alumno"
    }
    r = requests.post(f"{BASE_URL}/auth/register", json=register_data)
    print_response("User Registration", r)
    if r.status_code != 201:
        print("⚠️  User might already exist, continuing with login...")
    
    # 3. Login
    print("\n3️⃣  Testing Login...")
    login_data = {
        "email": "testuser@pucp.edu.pe",
        "password": "password123"
    }
    r = requests.post(f"{BASE_URL}/auth/login", json=login_data)
    print_response("Login", r)
    assert r.status_code == 200, "Login failed"
    
    token = r.json()["data"]["access_token"]
    headers = {"Authorization": f"Bearer {token}"}
    print(f"✅ Token obtained: {token[:50]}...")
    
    # 4. Get current user
    print("\n4️⃣  Testing Get Current User (/auth/me)...")
    r = requests.get(f"{BASE_URL}/auth/me", headers=headers)
    print_response("Get Current User", r)
    assert r.status_code == 200, "Get current user failed"
    
    current_user = r.json()["data"]
    user_id = current_user["id"]
    
    # 5. List users (as alumno - should see only themselves)
    print("\n5️⃣  Testing List Users (as alumno)...")
    r = requests.get(f"{BASE_URL}/users?page=1&size=10", headers=headers)
    print_response("List Users (alumno)", r)
    assert r.status_code == 200, "List users failed"
    
    # 6. Get specific user (themselves)
    print("\n6️⃣  Testing Get User by ID (self)...")
    r = requests.get(f"{BASE_URL}/users/{user_id}", headers=headers)
    print_response(f"Get User {user_id}", r)
    assert r.status_code == 200, "Get user by ID failed"
    
    # 7. Try to access another user (should fail for alumno)
    print("\n7️⃣  Testing Access to Another User (should fail)...")
    r = requests.get(f"{BASE_URL}/users/1", headers=headers)
    print_response("Get User 1 (should fail)", r)
    if r.status_code == 403:
        print("✅ Correctly blocked - alumno cannot see other users")
    
    # 8. Update own profile
    print("\n8️⃣  Testing Update Own Profile...")
    update_data = {
        "full_name": "Usuario de Prueba Actualizado"
    }
    r = requests.put(f"{BASE_URL}/users/{user_id}", json=update_data, headers=headers)
    print_response("Update Own Profile", r)
    assert r.status_code == 200, "Update profile failed"
    
    # 9. Try to update role (should fail for alumno)
    print("\n9️⃣  Testing Update Role (should fail)...")
    update_data = {
        "role_id": 2
    }
    r = requests.put(f"{BASE_URL}/users/{user_id}", json=update_data, headers=headers)
    print_response("Update Role (should fail)", r)
    if r.status_code == 403:
        print("✅ Correctly blocked - alumno cannot change role")
    
    # 10. Try to create user (should fail - only superadmin)
    print("\n🔟 Testing Create User (should fail)...")
    create_data = {
        "email": "newuser@pucp.edu.pe",
        "password": "password123",
        "full_name": "New User",
        "role_id": 3,
        "status": "active"
    }
    r = requests.post(f"{BASE_URL}/users", json=create_data, headers=headers)
    print_response("Create User (should fail)", r)
    if r.status_code == 403:
        print("✅ Correctly blocked - only superadmin can create users")
    
    # 11. Test with invalid token
    print("\n1️⃣1️⃣  Testing Invalid Token...")
    bad_headers = {"Authorization": "Bearer invalid_token"}
    r = requests.get(f"{BASE_URL}/auth/me", headers=bad_headers)
    print_response("Invalid Token (should fail)", r)
    if r.status_code == 401:
        print("✅ Correctly rejected invalid token")
    
    print("\n" + "="*60)
    print("✨ Test Suite Completed!")
    print("="*60)
    print("\n📊 Summary:")
    print("  ✅ Health check working")
    print("  ✅ User registration working")
    print("  ✅ Login and JWT working")
    print("  ✅ Get current user working")
    print("  ✅ List users with RBAC working")
    print("  ✅ Get user by ID with RBAC working")
    print("  ✅ Update profile working")
    print("  ✅ RBAC permissions enforced")
    print("\n🎉 All tests passed! API is working correctly.")


if __name__ == "__main__":
    try:
        test_api()
    except requests.exceptions.ConnectionError:
        print("\n❌ ERROR: Cannot connect to API at " + BASE_URL)
        print("Make sure the server is running with:")
        print("  python -m app.main")
        print("  or")
        print("  uvicorn app.main:app --reload --port 8000")
    except AssertionError as e:
        print(f"\n❌ TEST FAILED: {e}")
    except Exception as e:
        print(f"\n❌ UNEXPECTED ERROR: {e}")
        import traceback
        traceback.print_exc()
